Rename files J2 plasticity CL
