+++ applications/StructuralMechanicsApplication/custom_constitutive/small_strain_j2_plasticity_plane_strain_2d.cpp
@@ -17,7 +17,7 @@
 // External includes
 
 // Project includes
-#include "linear_j2_plasticity_plane_strain_2d.h"
+#include "small_strain_j2_plasticity_plane_strain_2d.h"
 #include "structural_mechanics_application_variables.h"
 
 namespace Kratos
