+++ applications/StructuralMechanicsApplication/custom_constitutive/small_strain_j2_plasticity_plane_strain_2d.h
@@ -19,7 +19,7 @@
 // External includes
 
 // Project includes
-#include "linear_j2_plasticity_3d.h"
+#include "small_strain_j2_plasticity_3d.h"
 
 namespace Kratos
 {
