+++ applications/StructuralMechanicsApplication/structural_mechanics_application.h
@@ -95,8 +95,8 @@
 #include "custom_constitutive/hyper_elastic_isotropic_neo_hookean_3d.h"
 #include "custom_constitutive/hyper_elastic_isotropic_neo_hookean_plane_strain_2d.h"
 #include "custom_constitutive/linear_elastic_orthotropic_2D_law.h"
-#include "custom_constitutive/linear_j2_plasticity_plane_strain_2d.h"
-#include "custom_constitutive/linear_j2_plasticity_3d.h"
+#include "custom_constitutive/small_strain_j2_plasticity_plane_strain_2d.h"
+#include "custom_constitutive/small_strain_j2_plasticity_3d.h"
 #include "custom_constitutive/linear_isotropic_damage_3D_law.h"
 #include "custom_constitutive/linear_isotropic_damage_plane_strain_2d.h"
